less random order in add_multi_benchmark()
Until now the order of inserted data points was
randomized due to their sorting into a hash for
entries of the same NAME and later using that
hash for the actual inserts. Hashes don't preserve
order, you know.

Now we store the order 'somewhat'. In particular
we remember the order in which the different NAMEs
appear and insert data points along that.

However, if multiple reports were processed, where each report
contains a metric A, B, and C, and we process 10 such reports at
once, then the we get first 10 different As in order, then 10
different B in order, then 10 different C in order, BUT NOT: the
1st A, 1st B, 1st C, then 2nd A, 2nd B, 2nd C, etc.

For our purposes this is still the right thing, as we
mostly care for the order inside a common NAME.

Order of incoming results is also preserved.